Extracted require_active_plugin decorator (#6801)
