+++ saleor/payment/gateways/razorpay/plugin.py
@@ -2,23 +2,13 @@ from typing import TYPE_CHECKING
 
 from saleor.plugins.base_plugin import BasePlugin, ConfigurationTypeField
 
-from ..utils import get_supported_currencies
+from ..utils import get_supported_currencies, require_active_plugin
 from . import GatewayConfig, capture, process_payment, refund
 
 GATEWAY_NAME = "Razorpay"
 
 if TYPE_CHECKING:
     from . import GatewayResponse, PaymentData
-
-
-def require_active_plugin(fn):
-    def wrapped(self, *args, **kwargs):
-        previous = kwargs.get("previous_value", None)
-        if not self.active:
-            return previous
-        return fn(self, *args, **kwargs)
-
-    return wrapped
 
 
 class RazorpayGatewayPlugin(BasePlugin):
